PerfectAttendance/courses_api/serializers.py
from rest_framework import serializers
from .models import Course


class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = ["title", "code", "building", "room"]
